+++ src/control/rotaryencoder.cpp
@@ -20,9 +20,36 @@
 // mt32-pi. If not, see <http://www.gnu.org/licenses/>.
 //
 
+#include <circle/timer.h>
+
 #include "control/rotaryencoder.h"
 
 // Based on encoder reading algorithm by Peter Dannegger: https://embdev.net/articles/Rotary_Encoders
+
+template <class T, T nMin, T nMax, size_t N>
+class QuadraticLookupTable
+{
+public:
+	constexpr QuadraticLookupTable()
+		: m_Coefficients{0}
+	{
+		constexpr float nScale = (static_cast<float>(nMax) - static_cast<float>(nMin)) / ((N - 1) * (N - 1));
+
+		for (size_t i = 0; i < N; ++i)
+			m_Coefficients[i] = static_cast<T>(nScale * i * i + nMin);
+	}
+
+	constexpr T operator[] (size_t nIndex) const { return m_Coefficients[nIndex]; }
+
+private:
+	T m_Coefficients[N];
+};
+
+// Delta-T threshold below which we begin accelerating
+constexpr u8 AccelThresholdMillis = 32;
+
+// Compile-time quadratic acceleration curve lookup table
+constexpr auto RotaryAccelLookupTable = QuadraticLookupTable<u8, 5, 16, AccelThresholdMillis>();
 
 CRotaryEncoder::CRotaryEncoder(TEncoderType Type, unsigned GPIOPinCLK, unsigned GPIOPinDAT)
 	: m_CLKPin(GPIOPinCLK, GPIOModeInputPullUp),
@@ -30,7 +57,9 @@ CRotaryEncoder::CRotaryEncoder(TEncoderType Type, unsigned GPIOPinCLK, unsigned 
 
 	  m_Type(Type),
 	  m_nDelta(0),
-	  m_nPreviousState(0)
+	  m_nPreviousState(0),
+
+	  m_nLastReadTime(0)
 {
 	if (m_CLKPin.Read() == LOW)
 		m_nPreviousState = 3;
@@ -61,6 +90,18 @@ s8 CRotaryEncoder::Read()
 			nResult  = m_nDelta;
 			m_nDelta = 0;
 			break;
+	}
+
+	// Apply acceleration curve
+	if (nResult != 0)
+	{
+		const unsigned int nTicks = CTimer::Get()->GetClockTicks();
+		const unsigned int nDeltaMillis = (nTicks - m_nLastReadTime) / 1000;
+
+		if (nDeltaMillis < AccelThresholdMillis)
+			nResult *= RotaryAccelLookupTable[nDeltaMillis];
+
+		m_nLastReadTime = nTicks;
 	}
 
 	return nResult;